+++ scalr/provider.go
@@ -88,6 +88,7 @@ func Provider() *schema.Provider {
 			"scalr_workspace_ids":            dataSourceScalrWorkspaceIDs(),
 			"scalr_workspaces":               dataSourceScalrWorkspaces(),
 			"scalr_event_bridge_integration": dataSourceScalrEventBridgeIntegration(),
+			"scalr_ssh_key":                  dataSourceScalrSSHKey(),
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
@@ -115,6 +116,7 @@ func Provider() *schema.Provider {
 			"scalr_workspace_run_schedule":         resourceScalrWorkspaceRunSchedule(),
 			"scalr_run_schedule_rule":              resourceScalrRunScheduleRule(),
 			"scalr_event_bridge_integration":       resourceScalrEventBridgeIntegration(),
+			"scalr_ssh_key":                        resourceScalrSSHKey(),
 		},
 
 		ConfigureContextFunc: providerConfigure,