+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **New data source:**  `scalr_module_versions` ([360](https://github.com/Scalr/terraform-provider-scalr/pull/360))
 - `data.scalr_variable`: new attributes `updated_at`, `updated_by` and `updated_by_email` ([354](https://github.com/Scalr/terraform-provider-scalr/pull/354))
 - `scalr_variable`: new attributes `updated_at`, `updated_by` and `updated_by_email` ([354](https://github.com/Scalr/terraform-provider-scalr/pull/354))
+- **New resource:** `scalr_ssh_key` ([#359](https://github.com/Scalr/terraform-provider-scalr/pull/359)
+- **New data source:** `scalr_ssh_key` ([#359](https://github.com/Scalr/terraform-provider-scalr/pull/359)
+- `scalr_workspace`: new attribute `ssh_key_id` ([#359](https://github.com/Scalr/terraform-provider-scalr/pull/359))
 
 ### Fixed